checkboxes converted into tiles
