+++ app/assets/stylesheets/checkboxes.scss
@@ -73,11 +73,11 @@
 }
 
 .selected{
-  background-color: $homepage-yellow
+  background-color: $medium-yellow
 }
 
 .unselected {
-  background-color: $medium-yellow
+  background-color: white
 }
 
 .timeslots-label {
