+++ app/assets/stylesheets/eventdetails.scss
@@ -14,9 +14,7 @@
 }
 
 .complete-schedule {
-  margin-top: 70px;
-  margin-right: 150px;
-  float: right;
+  margin-top: 35px;
   margin-bottom: 20px;
 }
