+++ app/assets/stylesheets/logins.scss
@@ -29,11 +29,3 @@
 .field {
   margin-bottom: 14px
 }
-
-// .line-it-up {
-//   display: flex;
-// }
-//
-// .label {
-//   margin-right: 10px
-// }
