+++ app/assets/stylesheets/text.scss
@@ -28,7 +28,7 @@ hr {
 
 .primary-subheader {
   color: $homepage-yellow;
-  font-size: 26px
+  font-size: 32px
 }
 
 .secondary-subheader {
