+++ app/javascript/react/components/NewEventForm.js
@@ -4,8 +4,7 @@ import humps from 'humps'
 import _ from 'lodash'
 import DatePicker from "react-datepicker"
 import ErrorList from "./ErrorList"
-import Timeslots from "../data/Timeslots.js"
-import CheckboxGroup from 'react-checkbox-group'
+import TimeslotTile from "./TimeslotTile.js"
 
 const NewEventForm = (props) => {
   const reset = {
@@ -107,6 +106,57 @@ const NewEventForm = (props) => {
     return <Redirect to="/" />
   }
 
+  const slots = [
+    '12:00-1:00am',
+    '1:00-2:00am',
+    '2:00-3:00am',
+    '3:00-4:00am',
+    '4:00-5:00am',
+    '5:00-6:00am',
+    '6:00-7:00am',
+    '7:00-8:00am',
+    '8:00-9:00am',
+    '9:00-10:00am',
+    '10:00-11:00am',
+    '11:00-12:00am',
+    '12:00am-1:00pm',
+    '1:00-2:00pm',
+    '2:00-3:00pm',
+    '3:00-4:00pm',
+    '4:00-5:00pm',
+    '5:00-6:00pm',
+    '6:00-7:00pm',
+    '7:00-8:00pm',
+    '8:00-9:00pm',
+    '9:00-10:00pm',
+    '10:00-11:00pm',
+    '11:00pm-12:00am'
+  ]
+
+  const handleTimeClick = (slot) => {
+    if (timeslots.includes(slot)) {
+      setTimeslots(timeslots.filter(timeslot => timeslot != slot))
+    } else {
+      setTimeslots([...timeslots, slot])
+    }
+  }
+
+  let times = slots.map((slot) => {
+    let selectionStatus = "unselected"
+    if (timeslots.includes(slot)) {
+      selectionStatus = "selected"
+    }
+
+    return(
+      <TimeslotTile
+        key={slot}
+        timeslot={slot}
+        selectionStatus={selectionStatus}
+        handleTimeClick={handleTimeClick}
+      />
+    )
+  })
+
   return(
     <div className="fading-background">
       <div className="ui stackable grid">
@@ -116,7 +166,7 @@ const NewEventForm = (props) => {
 
       <form onSubmit={handleEventSubmit}>
         <ErrorList errors={errors} />
-        <div className="eight wide column">
+        <div className="seven wide column">
         <div className="rsvp-form">
           <div className="form-field">
             <label> Event Window Name *
@@ -195,19 +245,19 @@ const NewEventForm = (props) => {
         </div>
       </div>
 
-      <div className="eight wide column">
+      <div className="seven wide column float right">
         <div className="timeslots">
-          <Timeslots
-            timeslots={timeslots}
-            setTimeslots={setTimeslots}
-          />
+          <div className="ui stackable grid">
+            {times}
+          </div>
           <div className="centered">
             <input className="form-button" type="submit" value="Send Invitations!"/>
           </div>
         </div>
       </div>
       </form>
-
+      <div className="spacer">
+      </div>
     </div>
     </div>
   )
