+++ app/javascript/react/components/TimeslotTile.js
@@ -2,13 +2,12 @@ import React from 'react';
 
 const TimeslotTile = (props) => {
   const handleClick = ((event) => {
-    props.handleTimeClick(event.currentTarget.id)
+    props.handleTimeClick(event.currentTarget.innerText)
   })
 
   return(
     <div>
-      <div className={`${props.selectionStatus} time centered four wide column`} onClick={handleClick} id={props.timeslot} name={props.timeslot}
-        value={props.timeslotId}>
+      <div className={`${props.selectionStatus} time centered three wide column`} onClick={handleClick} name={props.timeslot}>
         {props.timeslot}
       </div>
     </div>
